Minor refactor to profast tools org
