+++ CHANGELOG.md
@@ -2,10 +2,11 @@
 ## 0.5.x [TBD]
 - Updates models for NumPy version 2.4.0
 - Update test values for WOMBAT update to 0.13.0
-- Added standlone iron DRI and steel EAF performance and cost models
+- Added standalone iron DRI and steel EAF performance and cost models
 - Added capability to have transport models that require user input parameters
 - Add geologic hydrogen surface processing converter
 - Add baseclass for caching functionality
+- Minor reorg for profast tools
 
 ## 0.5.1 [December 18, 2025]
